🧹 Revert nginx resource
Remove the nginx resource added in #6783. The chunked binary scanning
improvement (scanBinaryForTag) and the apache2.conf parse guard are
retained in apache2.go since they are general improvements.
