+++ providers/os/resources/os.lr
@@ -862,77 +862,6 @@ private apache2.conf.directory @defaults("path") {
   params map[string]string
 }
 
-// Nginx HTTP Server
-nginx {
-  // Nginx version (e.g., "1.25.3")
-  version() string
-  // Compiled-in modules from configure arguments
-  modules() []string
-}
-
-// Nginx HTTP Server configuration
-nginx.conf {
-  init(path? string)
-  // Primary configuration file
-  file() file
-  // All configuration files (main + included fragments)
-  files(file) []file
-  // Flat key-value directives from main + http context
-  params(file) map[string]string
-  // User directive
-  user(params) string
-  // Worker processes setting
-  workerProcesses(params) string
-  // Error log path
-  errorLog(params) string
-  // Listen addresses/ports across all server blocks
-  listenAddresses(file) []string
-  // HTTP-level directives (inside http block, outside server blocks)
-  httpParams(file) map[string]string
-  // Server blocks
-  servers(file) []nginx.conf.server
-  // Upstream blocks
-  upstreams(file) []nginx.conf.upstream
-}
-
-// Nginx server block
-private nginx.conf.server @defaults("serverName") {
-  // Server name(s)
-  serverName string
-  // Listen directives (comma-separated if multiple)
-  listen string
-  // Document root
-  root string
-  // Whether SSL is configured
-  ssl bool
-  // Location blocks
-  locations []nginx.conf.location
-  // All directives within this server block
-  params map[string]string
-}
-
-// Nginx upstream block
-private nginx.conf.upstream @defaults("name") {
-  // Upstream group name
-  name string
-  // Backend server addresses
-  servers []string
-  // All directives within this upstream block
-  params map[string]string
-}
-
-// Nginx location block
-private nginx.conf.location @defaults("path") {
-  // Location path or pattern
-  path string
-  // Proxy pass target
-  proxyPass string
-  // Document root override
-  root string
-  // All directives within this location block
-  params map[string]string
-}
-
 // systemd journald configuration
 journald.config {
   init(path? string)
