+++ providers/os/resources/os.lr.go
@@ -80,11 +80,6 @@ const (
 	ResourceApache2ConfModule          string = "apache2.conf.module"
 	ResourceApache2ConfVirtualHost     string = "apache2.conf.virtualHost"
 	ResourceApache2ConfDirectory       string = "apache2.conf.directory"
-	ResourceNginx                      string = "nginx"
-	ResourceNginxConf                  string = "nginx.conf"
-	ResourceNginxConfServer            string = "nginx.conf.server"
-	ResourceNginxConfUpstream          string = "nginx.conf.upstream"
-	ResourceNginxConfLocation          string = "nginx.conf.location"
 	ResourceJournaldConfig             string = "journald.config"
 	ResourceJournaldConfigSection      string = "journald.config.section"
 	ResourceJournaldConfigSectionParam string = "journald.config.section.param"
@@ -459,26 +454,6 @@ func init() {
 		"apache2.conf.directory": {
 			// to override args, implement: initApache2ConfDirectory(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[string]*llx.RawData, plugin.Resource, error)
 			Create: createApache2ConfDirectory,
-		},
-		"nginx": {
-			// to override args, implement: initNginx(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[string]*llx.RawData, plugin.Resource, error)
-			Create: createNginx,
-		},
-		"nginx.conf": {
-			Init:   initNginxConf,
-			Create: createNginxConf,
-		},
-		"nginx.conf.server": {
-			// to override args, implement: initNginxConfServer(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[string]*llx.RawData, plugin.Resource, error)
-			Create: createNginxConfServer,
-		},
-		"nginx.conf.upstream": {
-			// to override args, implement: initNginxConfUpstream(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[string]*llx.RawData, plugin.Resource, error)
-			Create: createNginxConfUpstream,
-		},
-		"nginx.conf.location": {
-			// to override args, implement: initNginxConfLocation(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[string]*llx.RawData, plugin.Resource, error)
-			Create: createNginxConfLocation,
 		},
 		"journald.config": {
 			Init:   initJournaldConfig,
@@ -1888,81 +1863,6 @@ var getDataFields = map[string]func(r plugin.Resource) *plugin.DataRes{
 	},
 	"apache2.conf.directory.params": func(r plugin.Resource) *plugin.DataRes {
 		return (r.(*mqlApache2ConfDirectory).GetParams()).ToDataRes(types.Map(types.String, types.String))
-	},
-	"nginx.version": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginx).GetVersion()).ToDataRes(types.String)
-	},
-	"nginx.modules": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginx).GetModules()).ToDataRes(types.Array(types.String))
-	},
-	"nginx.conf.file": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConf).GetFile()).ToDataRes(types.Resource("file"))
-	},
-	"nginx.conf.files": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConf).GetFiles()).ToDataRes(types.Array(types.Resource("file")))
-	},
-	"nginx.conf.params": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConf).GetParams()).ToDataRes(types.Map(types.String, types.String))
-	},
-	"nginx.conf.user": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConf).GetUser()).ToDataRes(types.String)
-	},
-	"nginx.conf.workerProcesses": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConf).GetWorkerProcesses()).ToDataRes(types.String)
-	},
-	"nginx.conf.errorLog": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConf).GetErrorLog()).ToDataRes(types.String)
-	},
-	"nginx.conf.listenAddresses": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConf).GetListenAddresses()).ToDataRes(types.Array(types.String))
-	},
-	"nginx.conf.httpParams": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConf).GetHttpParams()).ToDataRes(types.Map(types.String, types.String))
-	},
-	"nginx.conf.servers": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConf).GetServers()).ToDataRes(types.Array(types.Resource("nginx.conf.server")))
-	},
-	"nginx.conf.upstreams": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConf).GetUpstreams()).ToDataRes(types.Array(types.Resource("nginx.conf.upstream")))
-	},
-	"nginx.conf.server.serverName": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfServer).GetServerName()).ToDataRes(types.String)
-	},
-	"nginx.conf.server.listen": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfServer).GetListen()).ToDataRes(types.String)
-	},
-	"nginx.conf.server.root": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfServer).GetRoot()).ToDataRes(types.String)
-	},
-	"nginx.conf.server.ssl": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfServer).GetSsl()).ToDataRes(types.Bool)
-	},
-	"nginx.conf.server.locations": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfServer).GetLocations()).ToDataRes(types.Array(types.Resource("nginx.conf.location")))
-	},
-	"nginx.conf.server.params": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfServer).GetParams()).ToDataRes(types.Map(types.String, types.String))
-	},
-	"nginx.conf.upstream.name": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfUpstream).GetName()).ToDataRes(types.String)
-	},
-	"nginx.conf.upstream.servers": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfUpstream).GetServers()).ToDataRes(types.Array(types.String))
-	},
-	"nginx.conf.upstream.params": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfUpstream).GetParams()).ToDataRes(types.Map(types.String, types.String))
-	},
-	"nginx.conf.location.path": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfLocation).GetPath()).ToDataRes(types.String)
-	},
-	"nginx.conf.location.proxyPass": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfLocation).GetProxyPass()).ToDataRes(types.String)
-	},
-	"nginx.conf.location.root": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfLocation).GetRoot()).ToDataRes(types.String)
-	},
-	"nginx.conf.location.params": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlNginxConfLocation).GetParams()).ToDataRes(types.Map(types.String, types.String))
 	},
 	"journald.config.file": func(r plugin.Resource) *plugin.DataRes {
 		return (r.(*mqlJournaldConfig).GetFile()).ToDataRes(types.Resource("file"))
@@ -5255,126 +5155,6 @@ var setDataFields = map[string]func(r plugin.Resource, v *llx.RawData) bool{
 	},
 	"apache2.conf.directory.params": func(r plugin.Resource, v *llx.RawData) (ok bool) {
 		r.(*mqlApache2ConfDirectory).Params, ok = plugin.RawToTValue[map[string]any](v.Value, v.Error)
-		return
-	},
-	"nginx.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginx).__id, ok = v.Value.(string)
-		return
-	},
-	"nginx.version": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginx).Version, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"nginx.modules": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginx).Modules, ok = plugin.RawToTValue[[]any](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConf).__id, ok = v.Value.(string)
-		return
-	},
-	"nginx.conf.file": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConf).File, ok = plugin.RawToTValue[*mqlFile](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.files": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConf).Files, ok = plugin.RawToTValue[[]any](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.params": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConf).Params, ok = plugin.RawToTValue[map[string]any](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.user": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConf).User, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.workerProcesses": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConf).WorkerProcesses, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.errorLog": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConf).ErrorLog, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.listenAddresses": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConf).ListenAddresses, ok = plugin.RawToTValue[[]any](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.httpParams": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConf).HttpParams, ok = plugin.RawToTValue[map[string]any](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.servers": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConf).Servers, ok = plugin.RawToTValue[[]any](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.upstreams": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConf).Upstreams, ok = plugin.RawToTValue[[]any](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.server.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfServer).__id, ok = v.Value.(string)
-		return
-	},
-	"nginx.conf.server.serverName": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfServer).ServerName, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.server.listen": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfServer).Listen, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.server.root": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfServer).Root, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.server.ssl": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfServer).Ssl, ok = plugin.RawToTValue[bool](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.server.locations": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfServer).Locations, ok = plugin.RawToTValue[[]any](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.server.params": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfServer).Params, ok = plugin.RawToTValue[map[string]any](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.upstream.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfUpstream).__id, ok = v.Value.(string)
-		return
-	},
-	"nginx.conf.upstream.name": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfUpstream).Name, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.upstream.servers": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfUpstream).Servers, ok = plugin.RawToTValue[[]any](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.upstream.params": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfUpstream).Params, ok = plugin.RawToTValue[map[string]any](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.location.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfLocation).__id, ok = v.Value.(string)
-		return
-	},
-	"nginx.conf.location.path": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfLocation).Path, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.location.proxyPass": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfLocation).ProxyPass, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.location.root": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfLocation).Root, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"nginx.conf.location.params": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlNginxConfLocation).Params, ok = plugin.RawToTValue[map[string]any](v.Value, v.Error)
 		return
 	},
 	"journald.config.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
@@ -13567,440 +13347,6 @@ func (c *mqlApache2ConfDirectory) GetAllowOverride() *plugin.TValue[string] {
 }
 
 func (c *mqlApache2ConfDirectory) GetParams() *plugin.TValue[map[string]any] {
-	return &c.Params
-}
-
-// mqlNginx for the nginx resource
-type mqlNginx struct {
-	MqlRuntime *plugin.Runtime
-	__id       string
-	// optional: if you define mqlNginxInternal it will be used here
-	Version plugin.TValue[string]
-	Modules plugin.TValue[[]any]
-}
-
-// createNginx creates a new instance of this resource
-func createNginx(runtime *plugin.Runtime, args map[string]*llx.RawData) (plugin.Resource, error) {
-	res := &mqlNginx{
-		MqlRuntime: runtime,
-	}
-
-	err := SetAllData(res, args)
-	if err != nil {
-		return res, err
-	}
-
-	// to override __id implement: id() (string, error)
-
-	if runtime.HasRecording {
-		args, err = runtime.ResourceFromRecording("nginx", res.__id)
-		if err != nil || args == nil {
-			return res, err
-		}
-		return res, SetAllData(res, args)
-	}
-
-	return res, nil
-}
-
-func (c *mqlNginx) MqlName() string {
-	return "nginx"
-}
-
-func (c *mqlNginx) MqlID() string {
-	return c.__id
-}
-
-func (c *mqlNginx) GetVersion() *plugin.TValue[string] {
-	return plugin.GetOrCompute[string](&c.Version, func() (string, error) {
-		return c.version()
-	})
-}
-
-func (c *mqlNginx) GetModules() *plugin.TValue[[]any] {
-	return plugin.GetOrCompute[[]any](&c.Modules, func() ([]any, error) {
-		return c.modules()
-	})
-}
-
-// mqlNginxConf for the nginx.conf resource
-type mqlNginxConf struct {
-	MqlRuntime *plugin.Runtime
-	__id       string
-	mqlNginxConfInternal
-	File            plugin.TValue[*mqlFile]
-	Files           plugin.TValue[[]any]
-	Params          plugin.TValue[map[string]any]
-	User            plugin.TValue[string]
-	WorkerProcesses plugin.TValue[string]
-	ErrorLog        plugin.TValue[string]
-	ListenAddresses plugin.TValue[[]any]
-	HttpParams      plugin.TValue[map[string]any]
-	Servers         plugin.TValue[[]any]
-	Upstreams       plugin.TValue[[]any]
-}
-
-// createNginxConf creates a new instance of this resource
-func createNginxConf(runtime *plugin.Runtime, args map[string]*llx.RawData) (plugin.Resource, error) {
-	res := &mqlNginxConf{
-		MqlRuntime: runtime,
-	}
-
-	err := SetAllData(res, args)
-	if err != nil {
-		return res, err
-	}
-
-	if res.__id == "" {
-		res.__id, err = res.id()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if runtime.HasRecording {
-		args, err = runtime.ResourceFromRecording("nginx.conf", res.__id)
-		if err != nil || args == nil {
-			return res, err
-		}
-		return res, SetAllData(res, args)
-	}
-
-	return res, nil
-}
-
-func (c *mqlNginxConf) MqlName() string {
-	return "nginx.conf"
-}
-
-func (c *mqlNginxConf) MqlID() string {
-	return c.__id
-}
-
-func (c *mqlNginxConf) GetFile() *plugin.TValue[*mqlFile] {
-	return plugin.GetOrCompute[*mqlFile](&c.File, func() (*mqlFile, error) {
-		if c.MqlRuntime.HasRecording {
-			d, err := c.MqlRuntime.FieldResourceFromRecording("nginx.conf", c.__id, "file")
-			if err != nil {
-				return nil, err
-			}
-			if d != nil {
-				return d.Value.(*mqlFile), nil
-			}
-		}
-
-		return c.file()
-	})
-}
-
-func (c *mqlNginxConf) GetFiles() *plugin.TValue[[]any] {
-	return plugin.GetOrCompute[[]any](&c.Files, func() ([]any, error) {
-		if c.MqlRuntime.HasRecording {
-			d, err := c.MqlRuntime.FieldResourceFromRecording("nginx.conf", c.__id, "files")
-			if err != nil {
-				return nil, err
-			}
-			if d != nil {
-				return d.Value.([]any), nil
-			}
-		}
-
-		vargFile := c.GetFile()
-		if vargFile.Error != nil {
-			return nil, vargFile.Error
-		}
-
-		return c.files(vargFile.Data)
-	})
-}
-
-func (c *mqlNginxConf) GetParams() *plugin.TValue[map[string]any] {
-	return plugin.GetOrCompute[map[string]any](&c.Params, func() (map[string]any, error) {
-		vargFile := c.GetFile()
-		if vargFile.Error != nil {
-			return nil, vargFile.Error
-		}
-
-		return c.params(vargFile.Data)
-	})
-}
-
-func (c *mqlNginxConf) GetUser() *plugin.TValue[string] {
-	return plugin.GetOrCompute[string](&c.User, func() (string, error) {
-		vargParams := c.GetParams()
-		if vargParams.Error != nil {
-			return "", vargParams.Error
-		}
-
-		return c.user(vargParams.Data)
-	})
-}
-
-func (c *mqlNginxConf) GetWorkerProcesses() *plugin.TValue[string] {
-	return plugin.GetOrCompute[string](&c.WorkerProcesses, func() (string, error) {
-		vargParams := c.GetParams()
-		if vargParams.Error != nil {
-			return "", vargParams.Error
-		}
-
-		return c.workerProcesses(vargParams.Data)
-	})
-}
-
-func (c *mqlNginxConf) GetErrorLog() *plugin.TValue[string] {
-	return plugin.GetOrCompute[string](&c.ErrorLog, func() (string, error) {
-		vargParams := c.GetParams()
-		if vargParams.Error != nil {
-			return "", vargParams.Error
-		}
-
-		return c.errorLog(vargParams.Data)
-	})
-}
-
-func (c *mqlNginxConf) GetListenAddresses() *plugin.TValue[[]any] {
-	return plugin.GetOrCompute[[]any](&c.ListenAddresses, func() ([]any, error) {
-		vargFile := c.GetFile()
-		if vargFile.Error != nil {
-			return nil, vargFile.Error
-		}
-
-		return c.listenAddresses(vargFile.Data)
-	})
-}
-
-func (c *mqlNginxConf) GetHttpParams() *plugin.TValue[map[string]any] {
-	return plugin.GetOrCompute[map[string]any](&c.HttpParams, func() (map[string]any, error) {
-		vargFile := c.GetFile()
-		if vargFile.Error != nil {
-			return nil, vargFile.Error
-		}
-
-		return c.httpParams(vargFile.Data)
-	})
-}
-
-func (c *mqlNginxConf) GetServers() *plugin.TValue[[]any] {
-	return plugin.GetOrCompute[[]any](&c.Servers, func() ([]any, error) {
-		if c.MqlRuntime.HasRecording {
-			d, err := c.MqlRuntime.FieldResourceFromRecording("nginx.conf", c.__id, "servers")
-			if err != nil {
-				return nil, err
-			}
-			if d != nil {
-				return d.Value.([]any), nil
-			}
-		}
-
-		vargFile := c.GetFile()
-		if vargFile.Error != nil {
-			return nil, vargFile.Error
-		}
-
-		return c.servers(vargFile.Data)
-	})
-}
-
-func (c *mqlNginxConf) GetUpstreams() *plugin.TValue[[]any] {
-	return plugin.GetOrCompute[[]any](&c.Upstreams, func() ([]any, error) {
-		if c.MqlRuntime.HasRecording {
-			d, err := c.MqlRuntime.FieldResourceFromRecording("nginx.conf", c.__id, "upstreams")
-			if err != nil {
-				return nil, err
-			}
-			if d != nil {
-				return d.Value.([]any), nil
-			}
-		}
-
-		vargFile := c.GetFile()
-		if vargFile.Error != nil {
-			return nil, vargFile.Error
-		}
-
-		return c.upstreams(vargFile.Data)
-	})
-}
-
-// mqlNginxConfServer for the nginx.conf.server resource
-type mqlNginxConfServer struct {
-	MqlRuntime *plugin.Runtime
-	__id       string
-	// optional: if you define mqlNginxConfServerInternal it will be used here
-	ServerName plugin.TValue[string]
-	Listen     plugin.TValue[string]
-	Root       plugin.TValue[string]
-	Ssl        plugin.TValue[bool]
-	Locations  plugin.TValue[[]any]
-	Params     plugin.TValue[map[string]any]
-}
-
-// createNginxConfServer creates a new instance of this resource
-func createNginxConfServer(runtime *plugin.Runtime, args map[string]*llx.RawData) (plugin.Resource, error) {
-	res := &mqlNginxConfServer{
-		MqlRuntime: runtime,
-	}
-
-	err := SetAllData(res, args)
-	if err != nil {
-		return res, err
-	}
-
-	// to override __id implement: id() (string, error)
-
-	if runtime.HasRecording {
-		args, err = runtime.ResourceFromRecording("nginx.conf.server", res.__id)
-		if err != nil || args == nil {
-			return res, err
-		}
-		return res, SetAllData(res, args)
-	}
-
-	return res, nil
-}
-
-func (c *mqlNginxConfServer) MqlName() string {
-	return "nginx.conf.server"
-}
-
-func (c *mqlNginxConfServer) MqlID() string {
-	return c.__id
-}
-
-func (c *mqlNginxConfServer) GetServerName() *plugin.TValue[string] {
-	return &c.ServerName
-}
-
-func (c *mqlNginxConfServer) GetListen() *plugin.TValue[string] {
-	return &c.Listen
-}
-
-func (c *mqlNginxConfServer) GetRoot() *plugin.TValue[string] {
-	return &c.Root
-}
-
-func (c *mqlNginxConfServer) GetSsl() *plugin.TValue[bool] {
-	return &c.Ssl
-}
-
-func (c *mqlNginxConfServer) GetLocations() *plugin.TValue[[]any] {
-	return &c.Locations
-}
-
-func (c *mqlNginxConfServer) GetParams() *plugin.TValue[map[string]any] {
-	return &c.Params
-}
-
-// mqlNginxConfUpstream for the nginx.conf.upstream resource
-type mqlNginxConfUpstream struct {
-	MqlRuntime *plugin.Runtime
-	__id       string
-	// optional: if you define mqlNginxConfUpstreamInternal it will be used here
-	Name    plugin.TValue[string]
-	Servers plugin.TValue[[]any]
-	Params  plugin.TValue[map[string]any]
-}
-
-// createNginxConfUpstream creates a new instance of this resource
-func createNginxConfUpstream(runtime *plugin.Runtime, args map[string]*llx.RawData) (plugin.Resource, error) {
-	res := &mqlNginxConfUpstream{
-		MqlRuntime: runtime,
-	}
-
-	err := SetAllData(res, args)
-	if err != nil {
-		return res, err
-	}
-
-	// to override __id implement: id() (string, error)
-
-	if runtime.HasRecording {
-		args, err = runtime.ResourceFromRecording("nginx.conf.upstream", res.__id)
-		if err != nil || args == nil {
-			return res, err
-		}
-		return res, SetAllData(res, args)
-	}
-
-	return res, nil
-}
-
-func (c *mqlNginxConfUpstream) MqlName() string {
-	return "nginx.conf.upstream"
-}
-
-func (c *mqlNginxConfUpstream) MqlID() string {
-	return c.__id
-}
-
-func (c *mqlNginxConfUpstream) GetName() *plugin.TValue[string] {
-	return &c.Name
-}
-
-func (c *mqlNginxConfUpstream) GetServers() *plugin.TValue[[]any] {
-	return &c.Servers
-}
-
-func (c *mqlNginxConfUpstream) GetParams() *plugin.TValue[map[string]any] {
-	return &c.Params
-}
-
-// mqlNginxConfLocation for the nginx.conf.location resource
-type mqlNginxConfLocation struct {
-	MqlRuntime *plugin.Runtime
-	__id       string
-	// optional: if you define mqlNginxConfLocationInternal it will be used here
-	Path      plugin.TValue[string]
-	ProxyPass plugin.TValue[string]
-	Root      plugin.TValue[string]
-	Params    plugin.TValue[map[string]any]
-}
-
-// createNginxConfLocation creates a new instance of this resource
-func createNginxConfLocation(runtime *plugin.Runtime, args map[string]*llx.RawData) (plugin.Resource, error) {
-	res := &mqlNginxConfLocation{
-		MqlRuntime: runtime,
-	}
-
-	err := SetAllData(res, args)
-	if err != nil {
-		return res, err
-	}
-
-	// to override __id implement: id() (string, error)
-
-	if runtime.HasRecording {
-		args, err = runtime.ResourceFromRecording("nginx.conf.location", res.__id)
-		if err != nil || args == nil {
-			return res, err
-		}
-		return res, SetAllData(res, args)
-	}
-
-	return res, nil
-}
-
-func (c *mqlNginxConfLocation) MqlName() string {
-	return "nginx.conf.location"
-}
-
-func (c *mqlNginxConfLocation) MqlID() string {
-	return c.__id
-}
-
-func (c *mqlNginxConfLocation) GetPath() *plugin.TValue[string] {
-	return &c.Path
-}
-
-func (c *mqlNginxConfLocation) GetProxyPass() *plugin.TValue[string] {
-	return &c.ProxyPass
-}
-
-func (c *mqlNginxConfLocation) GetRoot() *plugin.TValue[string] {
-	return &c.Root
-}
-
-func (c *mqlNginxConfLocation) GetParams() *plugin.TValue[map[string]any] {
 	return &c.Params
 }
