+++ providers/os/resources/os.lr.versions
@@ -612,36 +612,6 @@ nftables.table.handle 11.8.14
 nftables.table.name 11.8.14
 nftables.table.rules 11.8.14
 nftables.tables 11.8.14
-nginx 11.8.14
-nginx.conf 11.8.14
-nginx.conf.errorLog 11.8.14
-nginx.conf.file 11.8.14
-nginx.conf.files 11.8.14
-nginx.conf.httpParams 11.8.14
-nginx.conf.listenAddresses 11.8.14
-nginx.conf.location 11.8.14
-nginx.conf.location.params 11.8.14
-nginx.conf.location.path 11.8.14
-nginx.conf.location.proxyPass 11.8.14
-nginx.conf.location.root 11.8.14
-nginx.conf.params 11.8.14
-nginx.conf.server 11.8.14
-nginx.conf.server.listen 11.8.14
-nginx.conf.server.locations 11.8.14
-nginx.conf.server.params 11.8.14
-nginx.conf.server.root 11.8.14
-nginx.conf.server.serverName 11.8.14
-nginx.conf.server.ssl 11.8.14
-nginx.conf.servers 11.8.14
-nginx.conf.upstream 11.8.14
-nginx.conf.upstream.name 11.8.14
-nginx.conf.upstream.params 11.8.14
-nginx.conf.upstream.servers 11.8.14
-nginx.conf.upstreams 11.8.14
-nginx.conf.user 11.8.14
-nginx.conf.workerProcesses 11.8.14
-nginx.modules 11.8.14
-nginx.version 11.8.14
 npm.package 10.2.1
 npm.package.cpes 10.2.1
 npm.package.files 10.2.1
